Adjust workflow for optional mouse compatibility
Fix #6

Include species and genome as config inputs and adapt downstream workflow to include this. Now for running scAbsloute on mice you should only need to change the species (to Mouse) and genome (to GRCm38) in the config file and run as normal with the snakemake input on your mice samples.

Adjusted some of the file save directories that did not match - these had caused errors in snakemake DAG creation.

Note: I could not get the density.smk job to run on the mice. It seems unnecessary (?) so I adapted the density.smk file to skip this analysis for mice and just make an empty output file "data/align/{wildcards.sample}.position.tsv".